sub team 1
I made added some changes and removed the header and footer so that it would be easier to merge. You just need to change the html contenet of interns page, request certificate page and take a skill page and update the style.css. I'm really soorry for the stress it looked different  from what I wanted to get